darwin-backend/pyproject.toml
[tool.poetry]
name = "darwin-ads-backend"
version = "0.1.0"
description = "FastAPI backend MVP for Darwin variants"
authors = ["Juan Diego Collazos <jucollaas@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
httpx = "^0.27.0"
sqlalchemy = { version = "^2.0.30", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100